# api/openapi/helm.openapi.yaml
openapi: "3.1.0"
info:
  title: HELM Kernel API
  version: 0.5.0
  description: |
    Deterministic execution kernel for AI tool calls.
    Drop-in OpenAI proxy + cryptographic receipts + offline-verifiable evidence packs.
  license:
    name: Apache-2.0
    url: https://github.com/Mindburn-Labs/helm-ai-kernel/blob/main/LICENSE
  contact:
    name: Mindburn Labs
    url: https://mindburn.org
servers:
  - url: http://localhost:8080
    description: Local development (Docker Compose)
  - url: http://localhost:9090
    description: Standalone proxy mode
...
paths:
...
  /api/v1/agent-ui/info:
    get:
      operationId: getAgentUIRuntimeInfo
      tags: [agent, console]
      summary: Read OSS Agent UI runtime capabilities
      security:
        - AdminBearerAuth: []
      responses:
        '200':
          description: AG-UI runtime metadata and read-only tool catalog
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                required: [runtime, protocol, transport, tools]
                properties:
                  runtime: { type: string }
                  protocol: { type: string, const: ag-ui }
                  transport: { type: string, const: sse }
                  tools:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
        '401':
          $ref: '#/components/responses/HelmError'
  /api/v1/agent-ui/run:
    post:
      operationId: runAgentUIRuntime
      tags: [agent, console]
      summary: Run the OSS read-only Agent UI assistant
      security:
        - AdminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              properties:
                threadId: { type: string }
                runId: { type: string }
                workspaceId: { type: string }
                currentSurface: { type: string }
                state:
                  type: object
                  additionalProperties: true
                messages:
                  type: array
                  items:
                    type: object
                    additionalProperties: true
                    properties:
                      id: { type: string }
                      role: { type: string }
                      content: { type: string }
      responses:
        '200':
          description: Server-sent AG-UI lifecycle event stream
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/HelmError'
        '401':
          $ref: '#/components/responses/HelmError'
  /api/ag-ui/info:
    get:
      operationId: getAGUIRuntimeInfoCompat
      tags: [agent, console]
      summary: Read OSS AG-UI runtime capabilities compatibility route
      security:
        - AdminBearerAuth: []
      responses:
        '200':
          description: AG-UI runtime metadata and read-only tool catalog
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                required: [runtime, protocol, transport, tools]
                properties:
                  runtime: { type: string }
                  protocol: { type: string, const: ag-ui }
                  transport: { type: string, const: sse }
                  tools:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
        '401':
          $ref: '#/components/responses/HelmError'
  /api/ag-ui/run:
    post:
      operationId: runAGUIRuntimeCompat
      tags: [agent, console]
      summary: Run the OSS read-only AG-UI assistant compatibility route
      security:
        - AdminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              properties:
                threadId: { type: string }
                runId: { type: string }
                workspaceId: { type: string }
                currentSurface: { type: string }
                state:
                  type: object
                  additionalProperties: true
                messages:
                  type: array
                  items:
                    type: object
                    additionalProperties: true
                    properties:
                      id: { type: string }
                      role: { type: string }
                      content: { type: string }
      responses:
        '200':
          description: Server-sent AG-UI lifecycle event stream
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/HelmError'
        '401':
          $ref: '#/components/responses/HelmError'
...
components:
  securitySchemes:
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: HELM_ADMIN_API_KEY
      description: |
        Self-hosted OSS admin bearer key. Set HELM_ADMIN_API_KEY on the server
        and send Authorization: Bearer <key> for protected runtime APIs.

  # ── Schemas ─────────────────────────────────────────
  schemas:
    HelmError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type, code, reason_code]
          properties:
            message:
              type: string
              description: Human-readable error message
            type:
              type: string
              enum: [invalid_request, authentication_error, permission_denied, not_found, internal_error]
            code:
              type: string
              description: Machine-readable error code
            reason_code:
              type: string
              description: HELM-specific reason code
              enum:
                - DENY_TOOL_NOT_FOUND
                - DENY_SCHEMA_MISMATCH
                - DENY_OUTPUT_DRIFT
                - DENY_BUDGET_EXCEEDED
                - DENY_APPROVAL_REQUIRED
                - DENY_APPROVAL_TIMEOUT
                - DENY_SANDBOX_TRAP
                - DENY_GAS_EXHAUSTION
                - DENY_TIME_LIMIT
                - DENY_MEMORY_LIMIT
                - DENY_POLICY_VIOLATION
                - DENY_TRUST_KEY_REVOKED
                - DENY_IDEMPOTENCY_DUPLICATE
                - ERROR_INTERNAL
            details:
              type: object
              description: Additional error context
              additionalProperties: true
...
  responses:
    HelmError:
      description: HELM error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HelmError'
          example:
            error:
              message: "Tool not found: unknown_tool"
              type: "invalid_request"
              code: "tool_not_found"
              reason_code: "DENY_TOOL_NOT_FOUND"